Add some missing std.
git-svn-id: https://svn.dealii.org/trunk@8514 0785d39b-7218-0410-832d-ea1e28bc413d

+++ tests/deal.II/support_point_map.cc
@@ -86,7 +86,7 @@ check ()
       GridGenerator::hyper_ball(tr, Point<dim>(), 1);
     }
   else
-    GridGenerator::hyper_cube(tr, -1./sqrt(dim),1./sqrt(dim));
+    GridGenerator::hyper_cube(tr, -1./std::sqrt(static_cast<double>(dim)),1./std::sqrt(static_cast<double>(dim)));
   if (dim != 1)
     {
       static const HyperBallBoundary<dim> boundary;
